+++ packages/react-dom/src/index.ts
@@ -4,5 +4,5 @@ export { FocusProps } from "./Focus";
 import Link from "./Link";
 import Focus from "./Focus";
 
-export * from "@curi/react";
+export * from "@curi/react-universal";
 export { Link, Focus };