+++ packages/react-dom/types/index.d.ts
@@ -2,5 +2,5 @@ export { LinkProps, NavigatingChildren } from "./Link";
 export { FocusProps } from "./Focus";
 import Link from "./Link";
 import Focus from "./Focus";
-export * from "@curi/react";
+export * from "@curi/react-universal";
 export { Link, Focus };