+++ packages/react-native/src/index.ts
@@ -1,6 +1,6 @@
 export { LinkProps, NavigatingChildren } from "./Link";
 
-export * from "@curi/react";
+export * from "@curi/react-universal";
 
 import Link from "./Link";