+++ packages/react-native/types/index.d.ts
@@ -1,4 +1,4 @@
 export { LinkProps, NavigatingChildren } from "./Link";
-export * from "@curi/react";
+export * from "@curi/react-universal";
 import Link from "./Link";
 export { Link };