+++ packages/react-universal/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* Migrated from `@curi/react` to `@curi/react-universal`.
 * Moved `react-dom` specific components (`<Link>` and `<Focus>`) to `@curi/react-dom`. `@curi/react-dom` also re-exports everything from `@curi/react`. You should never need to use `@curi/react` yourself, but instead choose the package for whichever renderer you are using `react-dom` or `react-native`.
 
 ## 1.0.0-beta.30